MMP-298, refactor standaloneEnvironment in order to use the MuleProcessController and make it copy the mule installation each time is created
MMP-298, keep refactoring StanAloneEnvironment and match StandAloneDeployment and AgentDeployment test in preparation for abstract class

MMP-298 change package name on IT

MMP-298, make all deployment test extend AbstractDeploymentTEst

MMP-298, final refactor
