+++ mule-artifact-it/mule-deployer-it/src/test/resources/empty-mule-deploy-agent-project/pom.xml
@@ -4,7 +4,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.mulesoft.muleesb.it</groupId>
-    <artifactId>agent</artifactId>
+    <artifactId>empty-mule-deploy-agent-project</artifactId>
     <version>1.0.0</version>
     <packaging>mule-application</packaging>
