+++ mule-artifact-it/mule-deployer-it/src/test/resources/empty-mule-deploy-standalone-project/pom.xml
@@ -4,7 +4,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>org.mule.tools.muleesb.it</groupId>
-    <artifactId>standalone</artifactId>
+    <artifactId>empty-mule-deploy-standalone-project</artifactId>
     <version>1.0.0</version>
     <packaging>mule-application</packaging>
