+++ mule-artifact-it/mule-packager-it/src/test/java/integration/test/mojo/SettingsConfigurator.java
@@ -8,7 +8,7 @@
  * LICENSE.txt file.
  */
 
-package integrationTests.mojo;
+package integration.test.mojo;
 
 import org.apache.maven.it.VerificationException;
 import org.apache.maven.it.Verifier;
@@ -16,12 +16,6 @@ import org.apache.maven.it.Verifier;
 import java.io.File;
 
 public interface SettingsConfigurator {
-
-  String DEFAULT_MULE_VERSION = "4.0.0";
-
-  String MAVEN_OPTS = "MAVEN_OPTS";
-  String MAVEN_OPTS_PROPERTY_KEY = "argLine";
-  String DEFAULT_DEBUG_ARG_LINE = "-agentlib:jdwp=transport=dt_socket,server=y,address=8000,suspend=y";
 
   default void setSettings(Verifier verifier) {
     String mavenSettings = System.getenv("MAVEN_SETTINGS");
@@ -38,9 +32,9 @@ public interface SettingsConfigurator {
   }
 
   default void setMavenOpts(Verifier verifier) {
-    String mavenOpts = System.getProperty(MAVEN_OPTS_PROPERTY_KEY);
+    String mavenOpts = System.getProperty("argLine");
     if (mavenOpts != null) {
-      verifier.setEnvironmentVariable(MAVEN_OPTS, mavenOpts);
+      verifier.setEnvironmentVariable("MAVEN_OPTS", mavenOpts);
     }
   }
