+++ mule-artifact-it/mule-packager-it/src/test/java/integration/test/mojo/ValidateMojoTest.java
@@ -7,7 +7,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package integrationTests.mojo;
+package integration.test.mojo;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.maven.it.VerificationException;
